+++ src/DirectedGraph/AbstractExpressionTree/index.ts
@@ -1,0 +1,1 @@
+export { AbstractExpressionTree } from "./AbstractExpressionTree";
