+++ src/DirectedGraph/AbstractObfuscatedExpressionTree/ObfuscatedError.ts
