+++ src/DirectedGraph/AbstractObfuscatedExpressionTree/index.ts
@@ -1,0 +1,1 @@
+export { AbstractObfuscatedExpressionTree } from "./AbstractObfuscatedExpressionTree";
