+++ src/DirectedGraph/AbstractObfuscatedExpressionTree/test-helpers/TestTreeVisitor.ts
@@ -1,5 +1,5 @@
-import { ITreeVisitor } from "../../DirectedGraph/ITree";
-import { TGenericNodeContent } from "../../DirectedGraph/types";
+import { ITreeVisitor } from "../../ITree";
+import { TGenericNodeContent } from "../../types";
 
 class TestTreeVisitor<T> implements ITreeVisitor<T> {
   public includeSubtrees = true;
