+++ src/DirectedGraph/keystore/index.ts
@@ -1,0 +1,1 @@
+export { KeyStore } from "./KeyStore";
